[fix] code refactoring

+++ Backend/crit/src/main/java/com/ssafy/crit/boards/service/BoardService.java
@@ -11,11 +11,12 @@ import com.ssafy.crit.boards.entity.Classification;
 import com.ssafy.crit.boards.entity.feeds.UploadFile;
 import com.ssafy.crit.boards.repository.UploadFileRepository;
 import com.ssafy.crit.boards.repository.ClassificationRepository;
-import com.ssafy.crit.boards.service.dto.BoardDto;
+import com.ssafy.crit.boards.service.dto.BoardResponseDto;
 import com.ssafy.crit.boards.entity.board.Board;
 import com.ssafy.crit.boards.repository.BoardRepository;
 
 import com.ssafy.crit.boards.service.dto.BoardSaveRequestDto;
+import com.ssafy.crit.boards.service.dto.BoardSaveResponseDto;
 import com.ssafy.crit.boards.service.dto.BoardShowSortDto;
 import com.ssafy.crit.common.s3.S3Uploader;
 
@@ -28,6 +29,9 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 
+/**
+ * author : 강민승
+ */
 @Service
 @RequiredArgsConstructor
 @Transactional
@@ -58,7 +62,7 @@ public class BoardService {
 
 	//개별 게시물 조회
 	@Transactional
-	public BoardDto getBoard(Long id) {
+	public BoardResponseDto getBoard(Long id) {
 		Board board = boardRepository.findById(id).orElseThrow(() -> {
 			return new IllegalArgumentException("Board Id를 찾을 수 없습니다.");
 		});
@@ -66,21 +70,27 @@ public class BoardService {
 		board.setViews(board.getViews() + 1);
 		boardRepository.save(board);
 
-		return BoardDto.toDto(board);
+		return BoardResponseDto.toDto(board);
 	}
 
-	public BoardSaveRequestDto write(List<MultipartFile> multipartFiles , BoardSaveRequestDto boardSaveRequestDto, User user) throws
+	public BoardSaveResponseDto write(List<MultipartFile> multipartFiles , BoardSaveRequestDto boardSaveRequestDto, User user) throws
 		IOException {
 
+		/**
+		 * 테스트 시 분류명이 없어도 새롭게 만들어주는 로직
+		 */
 		Classification classification = classificationRepository.findByCategory(boardSaveRequestDto.getClassification())
 			.orElseGet(() -> {
-				// Create and save a new Classification.
 				Classification newClassification = new Classification();
-				// Ensure that you're setting the Category here.
 				newClassification.setCategory(boardSaveRequestDto.getClassification());
 				classificationRepository.save(newClassification);
 				return newClassification;
 			});
+
+//		Classification classification = classificationRepository.findByCategory(boardSaveRequestDto.getClassification())
+//			.orElseThrow(() -> {
+//				return new IllegalArgumentException("일치하는 분류명이 없습니다.");
+//			});
 
 		List<String> storeFileResult = new ArrayList<>();
 		Board board = Board.builder()
@@ -110,21 +120,25 @@ public class BoardService {
 			}
 		}
 
-		boardSaveRequestDto.setId(board.getId());
-		boardSaveRequestDto.setImageFiles(storeFileResult);
-
-
-		return boardSaveRequestDto;
+		return BoardSaveResponseDto.builder()
+				.id(board.getId())
+				.title(board.getTitle())
+				.content(board.getContent())
+				.writer(user.getId())
+				.classification(classification.getCategory())
+				.imageFiles(storeFileResult)
+				.build();
 	}
 
 
-	public BoardDto update(Long id, BoardDto boardDto, List<MultipartFile> multipartFiles) throws IOException {
+	public BoardResponseDto update(Long id, BoardResponseDto boardDto, List<MultipartFile> multipartFiles, User user) throws IOException {
 		Board board = boardRepository.findById(id).orElseThrow(() -> {
 			return new IllegalArgumentException("Board Id를 찾을 수 없습니다!");
 		});
 
-		User user = board.getUser();
-
+		if(!board.getUser().getId().equals(user.getId())){
+			throw new IllegalArgumentException("글 수정권한이 없습니다.");
+		}
 		List<UploadFile> uploadFile = uploadFileRepository.findAllByBoardsId(id);
 
 		uploadFileRepository.deleteAll(uploadFile);
@@ -157,16 +171,20 @@ public class BoardService {
 
 		boardRepository.save(board);
 
-		return BoardDto.toDto(board);
+		return BoardResponseDto.toDto(board);
 	}
 
 	// 게시글 삭제
-	public void delete(Long id) {
+	public void delete(Long id, User user) {
 		// 매개변수 id를 기반으로, 게시글이 존재하는지 먼저 찾음
 		// 게시글이 없으면 오류 처리
 		Board board = boardRepository.findById(id).orElseThrow(() -> {
 			return new IllegalArgumentException("Board Id를 찾을 수 없습니다!");
 		});
+
+		if(!board.getUser().getId().equals(user.getId())){
+			throw new IllegalArgumentException("글 삭제권한이 없습니다.");
+		}
 		// 게시글이 있는 경우 삭제처리
 		boardRepository.deleteById(id);
 	}
